Introduce ITaskParameterLoggingOptions.
Tasks can customize the parameter logging behavior. All new API is internal.
