+++ ref/Microsoft.Build.Tasks.Core/net/Microsoft.Build.Tasks.Core.cs
@@ -171,7 +171,6 @@ namespace Microsoft.Build.Tasks
         [Microsoft.Build.Framework.RequiredAttribute]
         public Microsoft.Build.Framework.ITaskItem[] Paths { get { throw null; } set { } }
         public override bool Execute() { throw null; }
-        protected override Microsoft.Build.Utilities.Task.ParameterLoggingOptions GetParameterLoggingOptions(string parameterName) { throw null; }
     }
     public partial class Copy : Microsoft.Build.Tasks.TaskExtension, Microsoft.Build.Framework.ICancelableTask, Microsoft.Build.Framework.ITask
     {
@@ -192,7 +191,6 @@ namespace Microsoft.Build.Tasks
         public bool UseSymboliclinksIfPossible { get { throw null; } set { } }
         public void Cancel() { }
         public override bool Execute() { throw null; }
-        protected override Microsoft.Build.Utilities.Task.ParameterLoggingOptions GetParameterLoggingOptions(string parameterName) { throw null; }
     }
     public partial class CreateCSharpManifestResourceName : Microsoft.Build.Tasks.CreateManifestResourceName
     {
@@ -385,7 +383,6 @@ namespace Microsoft.Build.Tasks
         public Microsoft.Build.Framework.ITaskItem Path { get { throw null; } set { } }
         public bool UpdateToAbsolutePaths { get { throw null; } set { } }
         public override bool Execute() { throw null; }
-        protected override Microsoft.Build.Utilities.Task.ParameterLoggingOptions GetParameterLoggingOptions(string parameterName) { throw null; }
     }
     public sealed partial class FormatUrl : Microsoft.Build.Tasks.TaskExtension
     {
@@ -558,7 +555,6 @@ namespace Microsoft.Build.Tasks
         public bool UsePreserializedResources { get { throw null; } set { } }
         public bool UseSourcePath { get { throw null; } set { } }
         public override bool Execute() { throw null; }
-        protected override Microsoft.Build.Utilities.Task.ParameterLoggingOptions GetParameterLoggingOptions(string parameterName) { throw null; }
     }
     public sealed partial class GenerateTrustInfo : Microsoft.Build.Tasks.TaskExtension
     {
@@ -719,7 +715,6 @@ namespace Microsoft.Build.Tasks
         [Microsoft.Build.Framework.RequiredAttribute]
         public Microsoft.Build.Framework.ITaskItem[] ItemsToHash { get { throw null; } set { } }
         public override bool Execute() { throw null; }
-        protected override Microsoft.Build.Utilities.Task.ParameterLoggingOptions GetParameterLoggingOptions(string parameterName) { throw null; }
     }
     [System.Runtime.InteropServices.GuidAttribute("00020401-0000-0000-C000-000000000046")]
     [System.Runtime.InteropServices.InterfaceTypeAttribute((System.Runtime.InteropServices.ComInterfaceType)(1))]
@@ -862,7 +857,6 @@ namespace Microsoft.Build.Tasks
         public bool HadAnyDuplicates { get { throw null; } set { } }
         public Microsoft.Build.Framework.ITaskItem[] Inputs { get { throw null; } set { } }
         public override bool Execute() { throw null; }
-        protected override Microsoft.Build.Utilities.Task.ParameterLoggingOptions GetParameterLoggingOptions(string parameterName) { throw null; }
     }
     public sealed partial class RequiresFramework35SP1Assembly : Microsoft.Build.Tasks.TaskExtension
     {
@@ -1286,7 +1280,6 @@ namespace Microsoft.Build.Tasks
         public bool Overwrite { get { throw null; } set { } }
         public bool WriteOnlyWhenDifferent { get { throw null; } set { } }
         public override bool Execute() { throw null; }
-        protected override Microsoft.Build.Utilities.Task.ParameterLoggingOptions GetParameterLoggingOptions(string parameterName) { throw null; }
     }
     public partial class XamlTaskFactory : Microsoft.Build.Framework.ITaskFactory
     {
