+++ src/Tasks/Hash.cs
@@ -37,7 +37,7 @@ namespace Microsoft.Build.Tasks
         [Output]
         public string HashResult { get; set; }
 
-        protected override ParameterLoggingOptions GetParameterLoggingOptions(string parameterName)
+        internal override ParameterLoggingOptions GetParameterLoggingOptions(string parameterName)
         {
             return parameterName switch
             {
